+++ src/Hacks/Frame Stepper/GameObjectHook.hpp
@@ -1,8 +1,0 @@
-/*#pragma once
-
-#include <Geode/Geode.hpp>
-#include <Geode/modify/PlayLayer.hpp>
-#include <Geode/modify/GJBaseGameLayer.hpp>
-#include "../../Client/Client.h"
-
-using namespace geode::prelude;*/